feat: create folders structure and navigation configure
