+++ Diaristando/App.tsx
@@ -1,20 +1,5 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
+import { Routes } from './src/navigation/routes';
 
 export default function App() {
-  return (
-    <View style={styles.container}>
-      <Text>Diaristando App!</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
+  return <Routes />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
