+++ Diaristando/src/navigation/routes.tsx
@@ -1,0 +1,10 @@
+import { NavigationContainer } from '@react-navigation/native';
+import { TabRoutes } from './tabRoutes';
+
+export function Routes() {
+  return (
+    <NavigationContainer>
+      <TabRoutes />
+    </NavigationContainer>
+  );
+}
