+++ Diaristando/src/screens/Inicio.tsx
@@ -1,0 +1,18 @@
+import { StyleSheet, Text, View } from 'react-native';
+
+export function Inicio() {
+  return (
+    <View style={styles.container}>
+      <Text>Diaristando App!</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
